Fixed class errors
