+++ src/main/java/com/mycompany/seniorproject/games/snake/SnakeGame.java
@@ -1,0 +1,264 @@
+
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/javafx/FXMLController.java to edit this template
+ */
+package com.mycompany.seniorproject.games.snake;
+
+import com.mycompany.seniorproject.App;
+import javafx.application.Application;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
+import javafx.stage.Stage;
+
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
+import javafx.scene.Group;
+import javafx.scene.canvas.Canvas;
+import javafx.scene.canvas.GraphicsContext;
+
+import javafx.scene.image.Image;
+import javafx.scene.paint.Color;
+import java.awt.Point;
+import javafx.animation.Animation;
+import javafx.animation.KeyFrame;
+import javafx.animation.Timeline;
+import static javafx.application.Application.launch;
+import javafx.event.EventHandler;
+//import java.awt.event.KeyEvent;
+import javafx.scene.input.KeyEvent;
+import javafx.scene.input.KeyCode;
+import javafx.scene.text.Font;
+import javafx.util.Duration;
+
+/**
+ * JavaFX App
+ */
+public class SnakeGame {
+
+    /*
+    Create a board for background
+    */
+    
+    private static final int WIDTH = 800;
+    private static final int HEIGHT = WIDTH;
+    private static final int ROWS = 20;
+    private static final int COLUMNS = ROWS;
+    private static final int SQUARE_SIZE = WIDTH / ROWS;
+    private static final String[] FOOD_IMAGE = new String[]{"img/ic_orange.png", "img/ic_watermelon.png",
+   "img/ic_apple.png" , "img/ic_berry.png", "img/ic_cherry.png", "img/ic_coconut.png", "img/ic_peach.png",
+    "img/ic_pomegranate.png", "img/ic_tomato.png"}; // String array to store the food images
+
+    private static final int RIGHT = 0;
+    private static final int LEFT = 1;
+    private static final int UP = 2;
+    private static final int DOWN = 3;
+
+    private GraphicsContext gc;
+    private List<Point> snakeBody = new ArrayList();
+    private Point snakeHead;
+    private Image foodImage;
+    private int foodX;
+    private int foodY;
+    private boolean gameOver;
+    private int currentDirection;
+    private int score = 0;
+    
+
+    public SnakeGame(Canvas c) {
+        
+        setGraphicsContext(c);
+
+        App.getScene().setOnKeyPressed(new EventHandler<KeyEvent>() {
+            @Override
+            public void handle(KeyEvent event) {
+                KeyCode code = event.getCode();
+                if (code == KeyCode.RIGHT || code == KeyCode.D) {
+                    if (currentDirection != LEFT) {
+                        currentDirection = RIGHT;
+                    }
+                } else if (code == KeyCode.LEFT || code == KeyCode.A) {
+                    if (currentDirection != RIGHT) {
+                    }
+                    currentDirection = LEFT;
+
+                } else if (code == KeyCode.UP || code == KeyCode.W) {
+                    if (currentDirection != DOWN) {
+                    }
+                    currentDirection = UP;
+                } else if (code == KeyCode.DOWN || code == KeyCode.C) {
+                    if (currentDirection != UP) {
+                    }
+                    currentDirection = DOWN;
+                }
+
+            }
+        });
+
+        for (int i = 0; i < 3; i++) { // for generate food
+
+            snakeBody.add(new Point(5, ROWS / 2));
+        }
+        snakeHead = snakeBody.get(0);
+
+        generateFood();
+
+        Timeline timeline = new Timeline(new KeyFrame(Duration.millis(130), e -> run(gc)));
+        timeline.setCycleCount(Animation.INDEFINITE);
+        timeline.play();
+    }
+    
+    private void setGraphicsContext(Canvas c) {
+        gc = c.getGraphicsContext2D();
+    }
+
+    private void run(GraphicsContext gc) {
+        if (gameOver) {
+            gc.setFill(Color.RED);
+            gc.setFont(new Font("Digital-7", 70));
+            gc.fillText("Game Over", WIDTH / 3.5, HEIGHT / 2);
+            return;
+        }
+        drawBackground(gc);
+        drawFood(gc);
+        drawSnake(gc);
+        drawScore();
+
+        for (int i = snakeBody.size() - 1; i >= 1; i--) {
+            snakeBody.get(i).x = snakeBody.get(i - 1).x;
+            snakeBody.get(i).y = snakeBody.get(i - 1).y;
+        }
+
+        switch (currentDirection) {
+            case RIGHT:
+                moveRight();
+                break;
+            case LEFT:
+                moveLeft();
+                break;
+            case UP:
+                moveUp();
+                break;
+            case DOWN:
+                moveDown();
+                break;
+
+        }
+        gameOver();
+        eatFood();
+    }
+
+    /*
+    draw the background using ROWS by COLUMNS that define on line 42,
+    Other color combination would be 
+    green - "AAS751", "A2d149" body: "4674E9"
+    rn: "WHEAT", "SKYBLUE", body: "NAVY"
+    */
+    private void drawBackground(GraphicsContext gc) {
+        for (int i = 0; i < ROWS; i++) {
+            for (int j = 0; j < COLUMNS; j++) {
+                if ((i + j) % 2 == 0) {
+                    gc.setFill(Color.web("WHEAT"));
+
+                } else {
+                    gc.setFill(Color.web("LIGHTBLUE"));
+                }
+                gc.fillRect(i * SQUARE_SIZE, j * SQUARE_SIZE, SQUARE_SIZE, SQUARE_SIZE);
+            }
+        }
+    }
+
+    /*
+    Using Math.random to generate food on the board
+    */
+    private void generateFood() {
+        start:
+        while (true) {
+            foodX = (int) (Math.random() * ROWS);
+            foodY = (int) (Math.random() * COLUMNS);
+
+            for (Point snake : snakeBody) {
+                if (snake.getX() == foodY && snake.getY() == foodY) {
+                    continue start;
+
+                }
+            }
+            foodImage = new Image(getClass().getResourceAsStream(FOOD_IMAGE[(int) (Math.random() * FOOD_IMAGE.length)]));
+            break;
+        }
+
+    }
+
+    private void drawFood(GraphicsContext gc) {
+        gc.drawImage(foodImage, foodX * SQUARE_SIZE, foodY * SQUARE_SIZE, SQUARE_SIZE, SQUARE_SIZE - 1);
+    }
+
+    private void drawSnake(GraphicsContext gc) {
+        gc.setFill(Color.web("NAVY"));
+
+        gc.fillRoundRect(snakeHead.getX() * SQUARE_SIZE, snakeHead.getY() * SQUARE_SIZE - 1, SQUARE_SIZE - 1,
+                SQUARE_SIZE - 1, 35, 35);
+
+        for (int i = 1; i < snakeBody.size(); i++) {
+            gc.fillRoundRect(snakeBody.get(i).getX() * SQUARE_SIZE, snakeBody.get(i).getY() * SQUARE_SIZE - 1,
+                    SQUARE_SIZE - 1, SQUARE_SIZE - 1, 20, 20);
+        }
+    }
+
+    private void moveRight() {
+        snakeHead.x++;
+    }
+
+    private void moveLeft() {
+        snakeHead.x--;
+    }
+
+    private void moveUp() {
+        snakeHead.y--;
+    }
+
+    private void moveDown() {
+        snakeHead.y++;
+    }
+
+    // if snake touches boarder, kills itself
+    public void gameOver() {
+        if (snakeHead.x < 0 || snakeHead.y < 0 || snakeHead.x * SQUARE_SIZE >= WIDTH
+                || snakeHead.y * SQUARE_SIZE >= HEIGHT) { // if snake touches border
+            gameOver = true;
+        }
+
+        //destroy itself
+        for (int i = 1; i < snakeBody.size(); i++) {
+            if (snakeHead.x == snakeBody.get(i).getX() && snakeHead.getY() == snakeBody.get(i).getY()) {
+                gameOver = true;
+                break;
+            }
+        }
+    }
+    
+    /*
+    If the snake Head touches the food point, socre + 5
+    */
+    private void eatFood(){
+        if (snakeHead.getX() == foodX && snakeHead.getY() == foodY){
+            snakeBody.add(new Point (-1, -1));
+            generateFood();
+            score += 5;
+        }
+    }
+    
+    /*
+    Text flowing on the board, displaying socre
+    */
+    private void drawScore(){
+        gc.setFill(Color.WHITE);
+        gc.setFont(new Font ("Digital-7", 35));
+        gc.fillText("Score: " + score, 10, 35);
+    }
+    
+  
+
+}
