+++ src/main/resources/com/mycompany/seniorproject/games/snake/SnakeGame.fxml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
+<?import javafx.scene.canvas.Canvas?>
 <?import javafx.scene.layout.BorderPane?>
-<?import javafx.scene.text.Text?>
 
-<BorderPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" stylesheets="@../../../styles/games/snakegame.css" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.mycompany.seniorproject.games.snake.SnakeGameController">
+<BorderPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" stylesheets="@../../../styles/games/snakegame.css" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.mycompany.seniorproject.games.snake.SnakeGameController">
    <center>
-      <Text onMouseClicked="#testClick" strokeType="OUTSIDE" strokeWidth="0.0" text="[Snake Game goes here]" BorderPane.alignment="CENTER" />
+      <Canvas fx:id="gameCanvas" height="800.0" width="800.0" BorderPane.alignment="CENTER" />
    </center>
 </BorderPane>
